Support ordered alternations (:ord).
Add a slight preference for the first element of each ordered
alternation. Ordered alternations are parsed into parser1 and parser2
where parser2 is essentially the rest of the parsers. So there is
implicit ordering when generating, but also add 1 to the first parser
of each pair so that even the left parser always gets place earlier in
the igen/freq generator which in particular will help with shrinking
to prioritize earlier elements.