Make nil edges of star/opt actual weighted paths.
Prior to this there was conflation of a star/opt main child with
0 weight as being equivalent to a removed star/opt node. But this
isn't actually true. When generating test cases, a star or opt with
0 weight for child 0 means that the node still exists but is
generating epsilon which may be useful case to test and or reduce
separate from star/opt having positive weight for child 0.

Or another way of looking at it: a star/opt node that isn't generating
content is still a valid node. It become invalid (removed) when it is
not generating content and not not generating content.