Fix reduce-weights no parents check.
Each removed node must have at least one weighted parent in nparents
that is not self-recursive (i.e. not lower than node in the same
rule). If there are none than this indicates that node has no parents
between itself and the root/start of the grammar so throw an exception
because the state of the weights/grammar would be invalid.