reduce-wtrek: reduced-subset param for efficiency.
The optional reduced-subset parameter allows the caller to say which
paths should be checked for reduce. If not specified then all weighted
nodes are searched. With the param things can be MUCH more efficient,
however, it can also mean the reduction may not be fully propogated if
the reduced-subset doesn't include all paths that have changed since
the last full reduce was run against a wtrek.