reduce: add prune-grammar->sorted-ebnf
Takes a grammar and optional wtrek and cycle-set. Converts the grammar
back to an ebnf representation. If wtrek is specified then any zero
weight branches of the grammar will be pruned out and not included in
the EBNF output. The rules in the EBNF are sorted in approximate
dependency order starting with the root/start node. If there are any
mutually recursive rules then cycle-set must be specified with a set
of rules to exclude from the sorting algorithm. These recursive rules
will be output first before the rest of the sorted rules.