[codegen] fix weight index of middle recur elements
For example, the following rule has a recursive element in the middle
of weighted children:

    r14 = '0' | '1' r14 | '2'

The recursive-gen generator results in an compound and simple
sub-generator. The compound generate previously had the correct weight
lookup however the simple generator previously had an off by one index
for the '2' branch (i.e. had an index of 1 instead of the correct 2).
This corrects the simple branch and also simplifies the path codegen
routes generator functions as part of the path refactoring.

This was discovered when a alternation path was chosen even though the
weight for that path was 0. This fixes that problem.